# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SANAD Studio API",
        "description": "API documentation for SANAD Studio Backend",
        "contact": {
            "email": "support@sanadstudio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pages/{section}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get page content for specific section",
                "operationId": "5b80109be737199a676f6f6b215461c3",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Page section name (hero, corporate_innovation, how_we_work, partner_with_us)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "hero",
                                "corporate_innovation",
                                "how_we_work",
                                "partner_with_us"
                            ]
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language code (en or ar)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page content retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "section": {
                                            "type": "string"
                                        },
                                        "language": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page section not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update page content for specific section and language",
                "operationId": "b700e31f76e511c272f28186fe73a229",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Page section name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "language",
                                    "data"
                                ],
                                "properties": {
                                    "language": {
                                        "type": "string",
                                        "enum": [
                                            "en",
                                            "ar"
                                        ],
                                        "example": "en"
                                    },
                                    "data": {
                                        "type": "object",
                                        "example": {
                                            "title": "Hero Section Title",
                                            "description": "Hero section description"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page content updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "section": {
                                            "type": "string"
                                        },
                                        "language": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/pages/{section}/both-languages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get page content for both languages",
                "operationId": "4682307fff5fc9a2d095d1bbd695e3ad",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Page section name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page content for both languages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "section": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "en": {
                                                    "type": "object"
                                                },
                                                "ar": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}